0.0.2 - Building out and decoupling functionality
### Added
- GitHub markdowns for issues/funding/pr templates/contributing/etc
- Importing rendering and rasterizing backend from ProtoTracer
    - Upper Camel Case files are unconverted
    - Empty files are in-progress
- Change Log

### Changed
- Decoupled functionality for Triangle 2D/3D geometry classes
- PlatformIO ini for support on embedded/non-embedded systems

+++ include/um3d/control/pid.hpp
@@ -12,6 +12,8 @@
 
 #pragma once
 
+#include <cstdint>
+
 /**
  * @class PID
  * @brief Implements a proportional-integral-derivative (PID) controller.
@@ -60,7 +62,7 @@ public:
      * @param currentMillis Current time in milliseconds.
      * @return Calculated control output.
      */
-    float Calculate(float setpoint, float processVariable, unsigned long currentMillis);
+    float Calculate(float setpoint, float processVariable, uint32_t currentMillis);
 
     /**
      * @brief Calculates the control output based on the setpoint and process variable.
